engine: gate the LTM table-index refusal on the enclosing freeze
Reviewer finding against `db6dc850`, whose refusal was keyed on the wrong thing. It
asked "is this a LOOKUP table argument?" when the question that matters is "is this
subtree already inside a freeze?" -- and those differ, because a `LOOKUP` can sit
INSIDE one of the other two pin-only descents:

    growth[Region] = pop[Region, young] + PREVIOUS(LOOKUP(pop[Region, idx], x))

The IR skips a table argument wherever it appears, so this reaches the same
no-occurrence path -- but the enclosing `PREVIOUS` lags everything inside it, index
reads included. Ceteris paribus already holds, and refusing dropped a perfectly
good per-element score plus every loop score through the edge. That is the same
over-decline `5ce2aa69` exists to undo, reintroduced one level down.

So the two loud verdicts are now distinguished, because only one of them depends on
the freeze. `IndexVerdict::Unspellable` -- another dimension's name, an axis this
target does not project, an over-arity index -- is a COMPILABILITY verdict: left
alone it keeps a dimension-name subscript that cannot resolve in a scalar fragment,
so it is loud everywhere, a freeze included. `IndexVerdict::RuntimeRead` -- a
variable or an expression selecting the element -- compiles fine, so it is purely a
CETERIS-PARIBUS verdict: loud in a bare table argument (which stays live), kept
inside a freeze (which lags it). The wrap already threads a `frozen` flag beside
`path` for exactly this kind of question, so the descent takes it too: `true` at the
`PREVIOUS`/`INIT` and whole-frozen-reducer sites, and the wrap's own `frozen` at the
table-argument site, which inherits the enclosing freeze and nothing more.

Naming the four possible verdicts is what makes the split legible; the previous
`Option<String>`-plus-bool encoding is what let two different reasons for declining
get conflated in the first place.

Three mutations, three distinct failures, no overlap: ignoring the freeze (always
decline) fails the frozen test, dropping the refusal entirely fails the bare test,
and gating `Unspellable` on the freeze fails the frozen test's second half. That
third probe initially passed -- nothing covered "unspellable inside a freeze is
still loud" -- so the frozen test carries both halves now.

The pin-descent tests move to their own `ltm_augment_pin_tests.rs`, mounted as a
child of the `tests` module so `use super::*` still resolves every helper. They had
pushed `ltm_augment_tests.rs` past the file cap, and they are a cohesive family:
one test per descent, each failing when its descent is deleted, plus the two
obligations (compile, stay ceteris-paribus) the file header now states.

All 16 char goldens byte-identical.
